Rename tokenTypes to userTypes
